refactor(web-analytics): isolate bot-tab state into botAnalyticsLogic
